Bailed on blockcypher.com
Changed to an object model that uses blockexplorer.com for BTC and explorer.litecoin.net for LTC because blockcypher.com limits the number of transactions per second so much that it caused errors refreshing my spreadsheets. The only drawback is that explorer.litecoin.net doesn't seem to have an API to obtain the number of transactions. I guess I can look into replacing the LTC explorer at some point.